requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
httpx==0.25.2
pydantic[email]==2.5.2
sqlalchemy==2.0.23
